@@ -9,11 +9,9 @@ class Menu {
 
   /// A `Map<String, Dish>`, where the String key is the dishID.
   /// Note that the key must be a string for `encode/decode` to work
-  Map<String, Dish> list;
+  late Map<String, Dish> list;
 
-  factory Menu() {
-    return _instance;
-  }
+  factory Menu() => _instance;
 
   Menu._internal() {
     load();
