@@ -1,8 +1,8 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:super_tooltip/super_tooltip.dart';
-import '../../generated/l10n.dart';
+// import 'package:super_tooltip/super_tooltip.dart';
+// import '../../generated/l10n.dart';
 import '../../theme/rally.dart';
 
 const _targetTime = Duration(seconds: 2);
@@ -14,7 +14,7 @@ const _targetTime = Duration(seconds: 2);
 class AnimatedLongClickableFAB extends StatefulWidget {
   final VoidCallback onLongPress;
 
-  const AnimatedLongClickableFAB({this.onLongPress});
+  const AnimatedLongClickableFAB({required this.onLongPress});
 
   @override
   _AnimatedLongClickableFABState createState() => _AnimatedLongClickableFABState();
@@ -22,16 +22,17 @@ class AnimatedLongClickableFAB extends StatefulWidget {
 
 class _AnimatedLongClickableFABState extends State<AnimatedLongClickableFAB>
     with TickerProviderStateMixin {
-  Timer t;
-  AnimationController valueController, colorController;
+  Timer? t;
+  late final AnimationController valueController, colorController;
   ColorTween cTween = ColorTween(begin: RallyColors.gray, end: RallyColors.primaryColor);
-  final tooltip = SuperTooltip(
-    popupDirection: TooltipDirection.up,
-    arrowTipDistance: 25.0,
-    borderWidth: 1.0,
-    backgroundColor: RallyColors.cardBackground,
-    content: Material(child: Text(S.current.lobby_tooltip, softWrap: true)),
-  );
+  // TODO update tooltip
+  // final tooltip = SuperTooltip(
+  //   popupDirection: TooltipDirection.up,
+  //   arrowTipDistance: 25.0,
+  //   borderWidth: 1.0,
+  //   backgroundColor: RallyColors.cardBackground,
+  //   content: Material(child: Text(AppLocalizations.of(context)!.lobby_tooltip, softWrap: true)),
+  // );
 
   _AnimatedLongClickableFABState() {
     valueController = AnimationController(
@@ -45,7 +46,7 @@ class _AnimatedLongClickableFABState extends State<AnimatedLongClickableFAB>
 
   @override
   void initState() {
-    WidgetsBinding.instance.addPostFrameCallback((_) => tooltip.show(context));
+    // WidgetsBinding.instance!.addPostFrameCallback((_) => tooltip.show(context));
     super.initState();
   }
 
@@ -86,12 +87,12 @@ class _AnimatedLongClickableFABState extends State<AnimatedLongClickableFAB>
               colorController.forward();
             },
             onTapUp: (_) {
-              if (t?.isActive ?? false) t.cancel();
+              if (t != null && t!.isActive) t!.cancel();
               valueController.reset();
               colorController.reset();
             },
             onTapCancel: () {
-              if (t?.isActive ?? false) t.cancel();
+              if (t != null && t!.isActive) t!.cancel();
               valueController.reset();
               colorController.reset();
             },
