@@ -1,17 +1,18 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 import '../../common/common.dart' show Money, MoneyFormatter, NumberEL100Formatter;
-import '../../generated/l10n.dart';
 import '../../theme/rally.dart';
 import '../../provider/src.dart' show Supplier, TableModel;
 
 class BottomNavBar extends StatelessWidget {
-  final String fromScreen, fromHeroTag;
+  final String fromScreen;
+  final String? fromHeroTag;
   final TableModel order;
 
-  BottomNavBar(this.order, {this.fromHeroTag, this.fromScreen});
+  BottomNavBar(this.order, {required this.fromScreen, this.fromHeroTag});
 
   @override
   Widget build(BuildContext context) {
@@ -41,10 +42,10 @@ class BottomNavBar extends StatelessWidget {
 
 class _CheckoutButton extends StatelessWidget {
   final TableModel order;
-  final String fromHeroTag;
+  final String? fromHeroTag;
   final String fromScreen;
 
-  _CheckoutButton(this.order, {this.fromHeroTag, @required this.fromScreen});
+  _CheckoutButton(this.order, {this.fromHeroTag, required this.fromScreen});
 
   @override
   Widget build(BuildContext context) {
@@ -60,8 +61,9 @@ class _CheckoutButton extends StatelessWidget {
           } else {
             final customerPaid = await _popUpPayment(context, order.totalPriceAfterDiscount);
             if (customerPaid != null) {
+              final s = await order.checkout(supplier: context.read<Supplier>());
               // ignore: unawaited_futures
-              order.checkout().then((o) => o.printReceipt(context, customerPaid));
+              order.printReceipt(context, customerPaid, s);
             }
           }
           Navigator.pop(context); // Go back to Lobby Screen
@@ -75,7 +77,7 @@ class _ApplyDiscountButton extends StatelessWidget {
   final TableModel order;
   final String fromScreen;
 
-  _ApplyDiscountButton(this.order, {@required this.fromScreen});
+  _ApplyDiscountButton(this.order, {required this.fromScreen});
 
   @override
   Widget build(BuildContext context) {
@@ -83,19 +85,17 @@ class _ApplyDiscountButton extends StatelessWidget {
       child: Icon(Icons.local_offer),
       minWidth: MediaQuery.of(context).size.width / 2,
       onPressed: () async {
-        final provider = Provider.of<Supplier>(context, listen: false);
-
         final discountPct = await _popUpDiscount(context, order.totalPricePreDiscount);
         if (discountPct != null) {
-          order.applyDiscount((100 - discountPct) / 100);
-          provider.notifyListeners();
+          final supplier = context.read<Supplier>();
+          order.applyDiscount((100 - discountPct) / 100, supplier);
         }
       },
     );
   }
 }
 
-Future<double> _popUpDiscount(BuildContext context, double totalPrice) {
+Future<double?> _popUpDiscount(BuildContext context, double totalPrice) {
   final notif = ValueNotifier('1'); // notify which type of TextField to display
   final percentageController = TextEditingController(text: '20');
   final fixedPriceController = TextEditingController(text: Money.format(10000));
@@ -105,17 +105,17 @@ Future<double> _popUpDiscount(BuildContext context, double totalPrice) {
     keyboardType: TextInputType.numberWithOptions(signed: true, decimal: true),
     inputFormatters: [NumberEL100Formatter()],
     textAlign: TextAlign.center,
-    decoration: InputDecoration(labelText: S.current.details_discount),
+    decoration: InputDecoration(labelText: AppLocalizations.of(context)!.details_discount),
   );
   final fixCtrl = TextField(
     controller: fixedPriceController,
     keyboardType: TextInputType.numberWithOptions(signed: true, decimal: false),
     inputFormatters: [MoneyFormatter()],
     textAlign: TextAlign.center,
-    decoration: InputDecoration(labelText: S.current.details_discount),
+    decoration: InputDecoration(labelText: AppLocalizations.of(context)!.details_discount),
   );
 
-  return showDialog<double>(
+  return showDialog<double?>(
     barrierDismissible: false,
     context: context,
     builder: (context) {
@@ -133,8 +133,8 @@ Future<double> _popUpDiscount(BuildContext context, double totalPrice) {
                 DropdownMenuItem(child: Center(child: Text('%')), value: '1'),
                 DropdownMenuItem(child: Center(child: Text(Money.symbol)), value: '2'),
               ],
-              onChanged: (v) {
-                notif.value = v;
+              onChanged: (String? v) {
+                if (v != null) notif.value = v;
               },
             );
             return Row(
@@ -161,7 +161,7 @@ Future<double> _popUpDiscount(BuildContext context, double totalPrice) {
                   // convert fixed price to percentage
                   discountPct = Money.unformat(selected.text) * 100 / totalPrice;
                 } else {
-                  discountPct = double.tryParse(selected.text);
+                  discountPct = double.parse(selected.text);
                 }
                 Navigator.pop<double>(context, discountPct);
               }
@@ -179,9 +179,9 @@ Future<double> _popUpDiscount(BuildContext context, double totalPrice) {
   );
 }
 
-Future<double> _popUpPayment(BuildContext scaffoldCtx, double needsToPay) {
+Future<double?> _popUpPayment(BuildContext scaffoldCtx, double needsToPay) {
   final t = TextEditingController(text: Money.format(needsToPay));
-  return showDialog<double>(
+  return showDialog<double?>(
     context: scaffoldCtx,
     barrierDismissible: false,
     builder: (context) {
@@ -192,7 +192,7 @@ Future<double> _popUpPayment(BuildContext scaffoldCtx, double needsToPay) {
           inputFormatters: [MoneyFormatter()],
           textAlign: TextAlign.center,
           decoration: InputDecoration(
-            labelText: S.current.details_customerPay,
+            labelText: AppLocalizations.of(context)!.details_customerPay,
             suffixText: Money.symbol,
           ),
         ),
@@ -202,10 +202,11 @@ Future<double> _popUpPayment(BuildContext scaffoldCtx, double needsToPay) {
             onPressed: () {
               final p = Money.unformat(t.text);
               if (p < needsToPay) {
-                final snackBar = SnackBar(content: Text(S.current.details_notEnough));
+                final snackBar =
+                    SnackBar(content: Text(AppLocalizations.of(context)!.details_notEnough));
                 ScaffoldMessenger.of(scaffoldCtx).showSnackBar(snackBar);
               } else {
-                Navigator.pop<double>(context, p);
+                Navigator.pop<double>(context, p.toDouble());
               }
             },
           ),
