@@ -7,10 +7,10 @@ import '../../provider/src.dart';
 
 class DetailsScreen extends StatelessWidget {
   final TableModel order;
-  final String fromHeroTag;
+  final String? fromHeroTag;
   final String fromScreen;
 
-  DetailsScreen(this.order, {this.fromHeroTag, this.fromScreen});
+  DetailsScreen(this.order, {this.fromHeroTag, required this.fromScreen});
 
   @override
   Widget build(BuildContext context) {
