@@ -4,22 +4,20 @@ import 'storage_engines/local_storage.dart';
 const appName = 'posapp';
 
 class DatabaseFactory {
-  static DatabaseFactory _singleton;
+  static final DatabaseFactory _singleton = DatabaseFactory._init();
 
   DatabaseFactory._init();
 
   /// Returns an [DatabaseConnectionInterface] instance, currently support `local-storage`
   factory DatabaseFactory() {
-    _singleton ??= DatabaseFactory._init();
-
     return _singleton;
   }
 
   DatabaseConnectionInterface create(
     String name, [
-    String path,
-    Map<String, dynamic> initialData,
-    String dbName,
+    String? path,
+    Map<String, dynamic>? initialData,
+    String? dbName,
   ]) {
     if (name == 'local-storage') {
       return LocalStorage(dbName ?? appName, path, initialData);
