@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 import '../../common/common.dart';
-import '../../generated/l10n.dart';
 import '../../provider/src.dart';
 import '../../storage_engines/connection_interface.dart';
 import '../popup_del.dart';
@@ -12,12 +12,11 @@ class OrderCard extends StatelessWidget {
   final DatabaseConnectionInterface storage;
   final Function(Order deletedOrder) onDeleted;
 
-  const OrderCard(this.order, this.storage, {this.onDeleted});
+  const OrderCard(this.order, this.storage, {required this.onDeleted});
 
   @override
   Widget build(BuildContext context) {
-    // in Dart, default is boolean is "null" !?
-    var del = order.isDeleted ?? false;
+    var del = order.isDeleted;
     return StatefulBuilder(
       builder: (context, setInternalState) {
         return Stack(
@@ -30,31 +29,32 @@ class OrderCard extends StatelessWidget {
                 child: ListTile(
                   leading: CircleAvatar(
                     child: Text(order.orderID.toString()),
-                    backgroundColor: del == true ? Colors.grey[400].withOpacity(0.5) : null,
+                    backgroundColor: del == true ? Colors.grey[400]!.withOpacity(0.5) : null,
                   ),
                   title: Text(
                     Common.extractYYYYMMDD3(order.checkoutTime),
-                    style: del == true ? TextStyle(color: Colors.grey[200].withOpacity(0.5)) : null,
+                    style:
+                        del == true ? TextStyle(color: Colors.grey[200]!.withOpacity(0.5)) : null,
                   ),
                   onLongPress: del == true
                       ? null
                       : () async {
                           var result = await popUpDelete(
                             context,
-                            title: Text(S.current.history_delPopUpTitle),
+                            title: Text(AppLocalizations.of(context)!.history_delPopUpTitle),
                           );
                           if (result == true) {
                             var ord = await storage.delete(order.checkoutTime, order.orderID);
                             setInternalState(() => del = ord.isDeleted);
-                            onDeleted?.call(ord);
+                            onDeleted.call(ord);
                           }
                         },
                   onTap: () {
                     Navigator.pushNamed(context, '/order-details', arguments: {
                       'state': TableModel(
-                        null,
                         order.tableID,
-                        StateObject.mock(
+                        TableState.mock(
+                          order.tableID,
                           order.lineItems,
                           orderID: order.orderID,
                           checkoutTime: order.checkoutTime,
@@ -68,7 +68,7 @@ class OrderCard extends StatelessWidget {
                     Money.format(order.totalPrice * order.discountRate),
                     style: TextStyle(
                       letterSpacing: 3,
-                      color: del == true ? Colors.grey[200].withOpacity(0.5) : Colors.lightGreen,
+                      color: del == true ? Colors.grey[200]!.withOpacity(0.5) : Colors.lightGreen,
                       fontSize: 20,
                     ),
                   ),
